Common: Convert remaining SFileInfo data members off of Windows-specific types

+++ Arc/Zlib.h
@@ -7,6 +7,6 @@ class CZlib
 public:
 	bool Decode(CArcFile* archive);
 	void DecompressFile(CArcFile* archive);
-	int Decompress(u8* dst, unsigned long* dst_size, const u8* src, unsigned long src_size);
-	int Decompress(u8* dst, unsigned long dst_size, const u8* src, unsigned long src_size);
+	int Decompress(u8* dst, u32* dst_size, const u8* src, u32 src_size);
+	int Decompress(u8* dst, u32 dst_size, const u8* src, u32 src_size);
 };
