+++ Extract/Circus.cpp
@@ -309,7 +309,7 @@ bool CCircus::DecodeCRX2(CArcFile* archive, const u8* src, size_t src_size)
 	std::vector<u8> pallet;
 	u8* pallet_ptr = nullptr;
 
-	unsigned long dst2_size = width * height * (bpp >> 3);
+	u32 dst2_size = width * height * (bpp >> 3);
 	std::vector<u8> dst2(dst2_size);
 
 	// Get palette
@@ -344,7 +344,7 @@ bool CCircus::DecodeCRX2(CArcFile* archive, const u8* src, size_t src_size)
 		// zlib decompression
 		CZlib zlib;
 
-		unsigned long dst_size = width * height * 4;
+		u32 dst_size = width * height * 4;
 		std::vector<u8> dst(dst_size);
 
 		while (true)
